:art: move and elements撲滅
