+++ src/stories/elements/Header.stories.ts
@@ -1,9 +1,9 @@
 import { Meta, StoryObj } from "@storybook/react"
 
-import { Header } from "@/components/presentational/elements/Header"
+import { Header } from "@/components/presentational/Header"
 
 export default {
-  title: "@/components/presentational/elements/Header",
+  title: "@/components/presentational/Header",
   component: Header,
 } satisfies Meta<typeof Header>
